kanban went to component
