+++ src/pages/student/KanbanBoard.scss
@@ -1,0 +1,59 @@
+.dashboard-app-tracker-content {
+  display: flex;
+  flex-direction: row;
+  margin-top: 1em;
+  justify-content: space-between;
+}
+
+.dashboard-app-tracker-content > *:not(:last-child) {
+  margin-right: 3vw
+}
+
+.dashboard-app-tracker-list {
+  flex: 1;
+  display:flex;
+  flex-direction: column;
+  flex-grow: 1;
+  min-width: 22vw;
+  min-height: 50px;
+}
+
+.dashboard-clubcard {
+  display: flex;
+  justify-content: space-between;
+  border-radius: 10px;
+  padding: 0;
+  box-shadow: 1px 1px 3px rgba(0, 0, 0, 0.3);
+  margin-bottom: 0.8em;
+  max-width: 400px;
+
+  button {
+      background: none;
+      border: none;
+  }
+
+  img{
+      width: 1.2em;
+      height: 1.2em;
+      margin: 0 0.5em;
+  }
+  
+  &:hover {
+      filter: brightness(85%);
+  }
+}
+
+.dashboard-clubcard-section-left > *, .dashboard-clubcard-section-right > * {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+}
+
+.dashboard-add-interested {
+  cursor: pointer;
+  background: none;
+  border: none;
+  font-weight: bold;
+  color: #8C8C8C;
+  font-family: Qanelas Soft;
+}
